refactor: split mcp operator tools
